split IPK code
- add rollup to bundle ipk
- need seperated tsconfigs to prevent unwanted files in dist folder
