+++ package.json
@@ -51,6 +51,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^10.0.1",
+    "@rollup/plugin-typescript": "^12.3.0",
     "@types/prompts": "^2.4.9",
     "@typescript-eslint/eslint-plugin": "^8.2.0",
     "@typescript-eslint/parser": "^8.2.0",
@@ -59,10 +60,16 @@
     "eslint-plugin-prefer-arrow": "^1.2.1",
     "eslint-plugin-prefer-arrow-functions": "^3.9.1",
     "eslint-plugin-unicorn": "^64.0.0",
+    "rollup": "^4.60.1",
+    "rollup-plugin-node-externals": "^8.1.2",
+    "rollup-plugin-terser": "^7.0.2",
     "typescript-eslint": "^8.58.0"
   },
   "scripts": {
-    "build": "tsc"
+    "prepack": "npm run build",
+    "build": "npm run build:plugin && npm run build:ipk",
+    "build:plugin": "tsc -d -p tsconfig_plugin.json",
+    "build:ipk": "rollup --config config/rollup.config.js"
   },
   "main": "dist/Plugin.js",
   "types": "dist/Plugin.d.ts"
