+++ tsconfig_plugin.json
@@ -10,7 +10,8 @@
     "types": ["node", "iitc"],
     "esModuleInterop": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
+    "declaration": true
   },
-  "include": ["./src"]
+  "files": ["./src/Plugin.ts"]
 }
